requirements.txt
absl-py==0.12.0
argcomplete==1.10.0
asgiref==3.3.4
astunparse==1.6.3
beautifulsoup4==4.8.0
cachetools==4.2.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
IMAPClient==2.1.0
importlib-metadata==4.0.1
joblib==1.0.1
Keras-Preprocessing==1.1.2
lxml==4.6.3
Markdown==3.3.4
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pdfminer.six==20181108
Pillow==8.2.0
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
python-pptx==0.6.18
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.22.2.post1
scipy==1.6.3
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
textract==1.6.3
tqdm==4.60.0
typing-extensions==3.7.4.3
tzlocal==1.5.1
urllib3==1.26.4
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.4.0
zipp==3.4.1
